flags for configuration

+++ image_service/resource.go
@@ -2,7 +2,7 @@
 package image_service
 
 type Config struct {
-	Quality  float32 `toml:"quality" comment:" Quality defines the quality of the stored webp Values from 1-100"`
+	Quality  float64 `toml:"quality" comment:" Quality defines the quality of the stored webp Values from 1-100"`
 	MaxWith  int     `toml:"width" comment:"defines the maximum with an image gets stored with"`
 	SavePath string  `toml:"save_path" comment:"specifies the directory where the image will be saved"`
 }
